+++ CHANGELOG.md
@@ -29,3 +29,5 @@
 ## 0.1.1-dev.1
 
 - Added Class diagrams.
+- Added ByteList's immutability tests.
+